⚡️ :: Add Button annotation
