+++ app/src/main/java/kr/hs/dgsw/smartschool/dui/screen/ButtonScreen.kt
@@ -25,8 +25,8 @@ import kr.hs.dgsw.smartschool.components.component.basic.button.ButtonType
 import kr.hs.dgsw.smartschool.components.component.basic.button.DodamLargeRoundedButton
 import kr.hs.dgsw.smartschool.components.component.basic.button.DodamMediumRoundedButton
 import kr.hs.dgsw.smartschool.components.component.basic.button.DodamSmallRoundedButton
-import kr.hs.dgsw.smartschool.components.component.basic.button.IconButton
-import kr.hs.dgsw.smartschool.components.component.basic.button.RadioButton
+import kr.hs.dgsw.smartschool.components.component.basic.button.DodamIconButton
+import kr.hs.dgsw.smartschool.components.component.basic.button.DodamRadioButton
 import kr.hs.dgsw.smartschool.components.component.set.appbar.DodamAppBar
 import kr.hs.dgsw.smartschool.components.theme.DodamColor
 import kr.hs.dgsw.smartschool.components.theme.IcAdd
@@ -97,9 +97,9 @@ fun ButtonRow(
     }
     Spacer(modifier = Modifier.height(20.dp))
     Row(verticalAlignment = Alignment.CenterVertically) {
-        IconButton(icon = { IcAdd(contentDescription = null) }, onClick = { onClick() }, type = type)
+        DodamIconButton(icon = { IcAdd(contentDescription = null) }, onClick = { onClick() }, type = type)
         Spacer(modifier = Modifier.width(20.dp))
-        RadioButton(selected = selected.value, onClick = { selected.value = !selected.value }, type = type)
+        DodamRadioButton(selected = selected.value, onClick = { selected.value = !selected.value }, type = type)
         Spacer(modifier = Modifier.width(20.dp))
         DodamSmallRoundedButton(onClick = { onClick() }, text = "Button", type = type)
         Spacer(modifier = Modifier.width(7.dp))
